T#804 P3-E: Extract files + upload module from server.ts
Sub-task E of T#786 Phase 3. Extract 7 file/upload handlers (~358 LoC) +
related file-handling constants (MAX_*, ALLOWED_EXTENSIONS, IMAGE_MAGIC) +
detectImageType helper into src/files/routes.ts.

## Changes

- New: src/files/routes.ts (379L) — registerFilesRoutes(app, sqlite, helpers)
  with FilesHelpers DI (hasSessionAuth, isTrustedRequest, isLocalNetwork,
  verifySessionToken, uploadsDir, sessionCookieName)
- server.ts: 3757L → 3399L (-358L)
- UPLOADS_DIR + ARCHIVE_DIR constants kept in server.ts (still used by guest
  avatar upload + telegram routes registration)

## Handlers
- POST /api/upload — multipart upload with magic-byte image validation +
  extension allowlist + double-extension reject + guest-image-only gate +
  sharp resize/EXIF strip + dual files+forum_attachments insert
- GET /api/files — paginated list with type/uploaded_by/context filters
- GET /api/files/stats — storage stats by type/context/archived
- GET /api/files/:id — file metadata (owner-only)
- GET /api/files/:id/download — owner-only download
- GET /api/f/:hash — hash-based download (local bypass + session/bearer for
  remote)
- DELETE /api/files/:id — soft delete (uploader or owner)

## Test plan
- [x] bun build clean (Norm #84 Layer 1)
- [x] Module registration before SPA fallback (T#783)
- [ ] Post-deploy smoke: GET /api/files no-auth → see if it lists (depends on
  endpoint auth pattern, may return 200 with empty list or 403)

Reviewer: @bertus per task assignment (security-touching — upload validation +
file serving).
